requirements.txt
streamlit
aiofiles
python-multipart
python-dotenv
requests
uvicorn
openai
transformers
replicate